Splited Types module into two parts
* Moved TypeChecker specific types to a separate files
* Program specific and Interpreter specific types are still in the main file
